feat: Implement HTML parsing and attribute extraction
- Added BasicAttributeExtractor, LinkAttributeExtractor, and InteractiveAttributeExtractor for extracting attributes from HTML elements.
- Introduced AttributeExtractorManager to manage multiple extractors and combine their results.
- Created constants for default parsing options and interactive tags.
- Developed parseHTMLToDOMMap and extractHTMLMetadata functions for parsing HTML strings into structured formats.
- Implemented structured parsing with parseHTMLToStructured function.
- Added utility functions for HTML validation, metadata extraction, and element visibility checks.
- Established pipelines for structured content and DOM map parsing.
- Defined types for parsed elements, DOM map nodes, and interactable elements.
- Enhanced error handling for parsing processes.
