+++ src/lib/html-parser/index.ts
@@ -1,0 +1,6 @@
+// Public exports - maintain backward compatibility
+export * from './types';
+export * from './constants';
+export * from './structured';
+export * from './dommap';
+export * from './interactable';
